samples/nucleus/requirements.txt
absl-py==0.8.1
alabaster==0.7.12
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.15
dataclasses==0.8
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
entrypoints==0.3
future==0.18.2
gast==0.3.2
geojson==2.5.0
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
hdf5storage==0.1.15
httplib2==0.9.2
idna==2.8
imageio==2.6.1
imagesize==1.1.0
imgaug==0.3.0
importlib-metadata==0.23
ipykernel==5.1.3
ipyparallel==6.2.4
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
netifaces==0.10.4
networkx==2.4
nose==1.3.7
notebook==6.0.2
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.1.26
opencv-python-headless==4.1.1.26
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.5
pyRFC3339==1.0
pyrsistent==0.15.5
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
scikit-image==0.16.2
scipy==1.3.3
scour==0.36
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
SimpleITK==1.2.4
simplejson==3.13.2
six==1.15.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.7.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xkit==0.0.0
zipp==0.6.0
zope.interface==4.3.2
